modifs for android (enable many panels/features which now compile but don't work)
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/perso/riviere/qroll/trunk@59332 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/libroll/somasound/somasoundprocessor.h
@@ -49,7 +49,7 @@ namespace soma
     virtual void process( int type );
     virtual unsigned inuse( int type )
     {
-      if( type >= _inuse.size() )
+      if( type >= (int) _inuse.size() )
         return false;
       return( _inuse[type] );
     }
